Speed up (process bits)
Use a temporary table instead of hash of hashes to store the results in the internal loop. The data and then copied to the final %test hash at the end of the method.

Also, use an array of indexes instead of a counter for the loop.

See: https://github.com/jherrero/eFORGE/commit/a54a23dab1a73d361ae335cae9881104648000b8